pyproject.toml
[tool.poetry]
name = "gleif_enricher"
version = "0.1.0"
description = "Enrichment tool for GLEIF data"
authors = ["Open Ownership <info@openownership.org>"]
license = "Apache-2.0"
packages = [
    { include = "gleif_enricher" }
]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.8"
elasticsearch = "^8.17.1"
requests = "^2.32.3"
boto3 = "^1.36.5"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gleif-enricher = "gleif_enricher.cli:main"
